Userpage: fix display-none on own profile in 8a400a209b4a.
With the changes in 8a400a209b4a, a bug was accidentally introduced
which resulted in one's own profile being wrapped in a `display: none;`
block due to a template conditional which caused the mobile
<div id="profile--actionbtns"> to not be closed when v.id != u.id
(cf. userpage.html circa L489).

Because the current conditional structure makes it hard to add an
analogous profile--actionbtns div to mobile, it has been removed.