Update loading.webp asset with improvements.
The original artist of the new loading animation provided a version
with one changed frame and a slower animation speed. Further, this
was scaled and compressed to about half the weight of the original
implementation.

This change is sufficiently minor that it's not worth cache-busting
for the minority of users who will have already cached the prior,
especially given it leaks into markup.