Fix hole_entered notif username.
Previously, the on_post_hole_entered notifs to hole followers would
contain the username of the admin who moved the post (due to
referencing the	`v` parameter of whomever performed the action which
placed the post in the hole). This has been corrected to use the post
author instead. Further, on_post_hole_entered only depends on a post
for behavior, though the optional `v` user is used if present.