Restyle glow award links for visibility.
At present, we typically distinguish links from other text by coloring
them, underlining on hover, and changing cursor to pointer on hover.
However, when a post is glow-awarded, both color and underline-on-hover
are suppressed, making links very hard to find in text. Given chiobu's
tendency to glow award effortposts, this can make posts hard to read.

My solution is to always show links as underlined in glow-awarded posts
and comments. Not terribly attached to this particular approach, but
it seems to undermine the award effect the least while largely
restoring accessibility of links.