Remove new sort on 'thread' in Submission.title.
Requested by multiple jannies. Rough timeline, as I understand it:
  - Circa 7mo ago, this logic was originally added for threads with
    'megathread' in the title.
  - Some time later, a checkbox on submission which sets the flag
    Submission.new does the same thing.
  - In af680d8a940e, change the check from 'megathread' to 'thread'.

There must've been some reason for the change of substring checked.
However, it routinely causes issues for the admins and confuses
users. Solution has been to retroactively update posts that currently
rely on the 'megathread' in title behavior to use the `new` flag and
to remove the logic going forward.