Implement hash-based cachebusting module.
Adds the files.helpers.assetcache module, which runs on worker startup
to compute CRC digests of files in given asset directories and provides
a Jinja filter to create cachebusted URLs based on those hashes.
This alleviates the need to increment cachebusting counters in query
strings entirely for files it is applied to.