Fix pruning of {SITE}_loggedin/out dicts.
For the past week, we noticed a gradual increase in CPU usage and
request times. Use of a sampling profiler revealed the time waas spent
in serializing/deserializing data stored in Redis. In particular, the
user counter dicts were filtered for calculation of the loggedin/out
counters, but the filtered versions were never stored.

To make concurrency safe, we still filter on every request, but at
least the resting data will eventually be appropriately filtered,
and this data is non-critical regardless.