Fix DM and modmail replies 404ing (#65)
messages got borked completely by this commit

this reverts commit b649a16b4612cda64db3fa8beccb6c2043bf084c.

---

Snakes note: DM and modmail replies 404'd because dispatched
to the wrong endpoint. The fix to that affected cancel buttons
in some contexts, so this PR also fixes those. Cancel button
presence and behavior should be same as it was a ~week ago
(i.e. correct), and all reply everywhere seem to work right.

Co-authored-by: justcool393 <justcool393@gmail.com>
Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/65
Co-authored-by: justcool393 <justcool393@noreply.fsdfsd.net>
Co-committed-by: justcool393 <justcool393@noreply.fsdfsd.net>