Fix env file references in startup scripts (#70)
Kindly,

I have downloaded this here repository again and when to trying to start the thing it did not work on my machine (TM)

I noticed that logs in docker compose complaining of multiple things about lacking environment variables

I added info to the readme.md about cp

This is inline with what is already written in the ubuntu setup script. So it should be good.

Now for the part in doubt ...

Then I noticed that there were two more references to "env" instead of ".env" in the other scripts and I corrected them too. Check if that is fine with you fine folk (TM)

Thanks,
granny

Co-authored-by: Your Name <madmarcolas@gmail.com>
Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/70
Co-authored-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>
Co-committed-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>