Stop skipping render of Follow/Unfollow button in user listing because profile is private (#94)
As per title "Stop skipping render of Follow/Unfollow button in user listing because profile is private"

Happens when browsing through the user listing (at least):
* /search/users
* /users

Seems strange to hide that button in the user listing and not the profile page itself. When profiles are private, there can be some interest in following/unfollowing even without clicking the profile. Tried to search git history but no additional sense was obtained

Co-authored-by: Your Name <madmarcolas@gmail.com>
Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/94
Co-authored-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>
Co-committed-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>