moved toggleelement to register_new_elements() (#111)
We currently cannot see (toggle the dropdown/button) reports that are hidden behind the "show more" (comments) button. This PR fixes that.

Example of what is happening without this PR:
1. open megathread. The thread is long includes hundreds of comments implying need to click "show more" (comments) button
2. without clicking any "show more" (comments) button, click the flaggers button on one of the first comments
3. flaggers are presented as expected
4. click the "show more" (comments) button at the end
5. click the flaggers button on one of the comments that was hidden behind the "show more" (comments) button
6. the button does nothing and has no event listenner associated

With this PR:
1. all flaggers buttons work regardless if hidden behind "Show More" (comments) or not. Tested up to two "Show More" (comments) buttons

I did the PR this way because the view_more button calls register_new_elements() and not bottom.js as a whole

There are other things using data-toggleelement like the undelete button on comments and the Preview button on /submit. Both of those still work as expected

Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/111
Co-authored-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>
Co-committed-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>