Post: Reduce amount of scrolling necessary after changing comments sorting (#114)
As per title. Especially useful on long posts both in mobile and desktop so that we don't have to scroll through the long post ater each change.

Don't have time for anything else today. Added to my list to analyze if we can add this skip in other pages especially those that have this same button

Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/114
Co-authored-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>
Co-committed-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>